refactor : 겹치는 UserController 파일을 하나로 합침
